try to solve get_ss3_exe() warning in tests

+++ tests/testthat/test-get_ss3_exe.R
@@ -1,6 +1,5 @@
 # example_path <- system.file("extdata", package = "r4ss")
 temp_path <- file.path(tempdir(), "test_ss3_exe")
-temp_path
 dir.create(temp_path, showWarnings = FALSE)
 # remove all artifacts created from testing. (developers: simply comment out
 # the line below if you want to keep artifacts for troubleshooting purposes)
@@ -11,7 +10,9 @@ test_that("executables are downloading default latest version", {
   download_filepath <- gsub(".*: ", "", download_loc)
   exe_name <- gsub(paste0(temp_path, "/"), "", download_filepath, fixed = TRUE)
   dir_temp <- file.path(temp_path, exe_name)
-  file.remove(download_filepath)
+  if (file.exists(download_filepath)) {
+    file.remove(download_filepath)
+  }
 
   expect_equal(dir_temp, download_filepath)
 })
